refactor: further simplify sync filenames
